fix include BTActionInitialization
